Refactor lower-generics pass into separate subpasses.
