+++ source/slang/slang-ir-insts.h
@@ -1477,6 +1477,11 @@ struct IRWitnessTable : IRInst
         return getOperand(0);
     }
 
+    void setConformanceType(IRInst* type)
+    {
+        setOperand(0, type);
+    }
+
     IR_LEAF_ISA(WitnessTable)
 };
