+++ source/slang/slang-ir-lower-generic-var.h
@@ -1,0 +1,12 @@
+// slang-ir-lower-generic-var.h
+#pragma once
+
+namespace Slang
+{
+    struct SharedGenericsLoweringContext;
+
+    /// Lower load and stores of generic local variables into raw pointer operations.
+    void lowerGenericVar(
+        SharedGenericsLoweringContext* sharedContext);
+
+}
